feat: Add sequencer implementation for Nonodo
The code changes include adding a new file `sequencer.go` in the `internal/rollup` package, which contains the implementation of two sequencer types: `InputBoxSequencer` and `EspressoSequencer`. The `InputBoxSequencer` implements the `Sequencer` interface and uses the `model.NonodoModel` to finish and get the next input. The `EspressoSequencer` also implements the `Sequencer` interface, but the implementation is not provided yet.

This commit adds the initial structure for dealing with sequencers in the Nonodo application, allowing the retrieval of the next input from either the InputBox or Espresso.

Refactor LegacyMode value to false in NewNonodoOpts function