feat: Add Fetch interface for rollup package
This commit adds a new file `fetcher.go` in the `internal/rollup` package, which contains the definition of the `Fetch` interface. The `Fetch` interface provides a method for fetching data using the `echo.Context` and an ID. This interface will be used to implement the fetching functionality in the Nonodo application.