feat: Add conditional check for existing sequencer in Register function
The code changes in this commit add a conditional check for the existing sequencer in the `Register` function of the `rollup.go` file. If the sequencer is not already set, a new `Sequencer` instance is created and set in the `NonodoModel`. This change ensures that the sequencer is only created if it doesn't already exist, preventing unnecessary re-creation of the sequencer.