feat: Update Celestia package to include RpcUrl parameter
The code changes in the Celestia package add a new RpcUrl parameter to the CelestiaOpts struct. This parameter allows users to specify a custom RPC URL for connecting to Celestia instead of using the default Anvil setup. The RpcUrl parameter is used in the addCelestiaSubcommands function to set the ethEndpointRPC variable. This update improves the flexibility and configurability of the Celestia command.