feat: Add chainId parameter to CallCelestiaRelay
The code changes in the `CallCelestiaRelay` function add a new `chainId` parameter. This parameter allows users to specify the chain ID of the network when making a relay call to Celestia. The `chainId` parameter is used in the function call in the `main.go` file. This update improves the flexibility and configurability of the Celestia command.