feat: Add support for sending files to Celestia Network
This commit introduces a new subcommand, `send-file`, to the Celestia command. It allows users to send a file to the Celestia Network. The file path and namespace are required parameters for this subcommand. This feature enhances the functionality of the Celestia command by enabling users to easily send files to the network.